requirements.txt
beautifulsoup4==4.10.0
django-apscheduler==0.6.0
django-cors-headers==3.8.0
django-dotenv==1.4.2
django==3.2.7
djangorestframework-simplejwt==4.8.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
requests==2.26.0
whitenoise==5.3.0
